server/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ofer-knowledge-rag"
version = "2025.1"
requires-python = ">=3.10"
dependencies = [
    "certifi == 2024.12.14",
    "Flask == 3.1.0",
    "flask-cors == 4.0.1",
    "flask_socketio == 5.0.0",
    "google-cloud-aiplatform == 1.77.0",
    "langchain-core == 0.3.29",
    "langchain-google-vertexai == 2.0.11",
    "langchain-text-splitters == 0.3.5",
    "langchain-community == 0.3.14 ",
    "lxml == 5.3.0",
    "matplotlib == 3.10.0",
    "nltk == 3.9.1",
    "numpy == 1.26.4",
    "opencv-python ==  4.10.0.84",
    "pathlib == 1.0.1",
    "PyPDF2 == 3.0.1",
    "python-docx == 1.1.2",
    "python-dotenv == 1.0.1",
    "requests == 2.32.3",
    "scikit-learn == 1.6.1",
    "sentence_transformers==3.2.1",
    "pdf2image==1.17.0",
    "pillow == 11.1.0",
    "pi_heif==0.21.0",
    "pypdf == 5.1.0",
    "pydantic == 2.9.2",
    "pytesseract == 0.3.13",
    "pdfminer.six == 20231228",
    "unstructured == 0.16.14",
    "unstructured_pytesseract == 0.3.13",
    "unstructured-inference == 0.8.2"
]
[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "ruff==0.8.6",
    "pre-commit==4.0.1"
]

[tool.setuptools.packages.find]
where = ["src"]
